fix security issues, unify configs, and improve error handling

+++ ProductService/src/main/java/org/example/productService/repository/ProductRepository.java
@@ -4,11 +4,14 @@ import org.example.productService.model.Product;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
 public interface ProductRepository extends MongoRepository<Product, String> {
     List<Product> findByName(String name);
-    List<Product> findByPriceLessThan(double price);
-    List<Product> findByPriceGreaterThan(double price);
+
+    List<Product> findByPriceLessThan(BigDecimal price);
+
+    List<Product> findByPriceGreaterThan(BigDecimal price);
 }
